5787 display coverage selection on review screen, refactor overview
